Don't load enviornment variables with xargs for dev
When the variables are loaded with xargs any values with spaces in them
get split.

This was causing the login to fail because authentication wasn't working
correctly.

In the server the enviorment is already loaded with `dotenv`. Use
`webpack-dotenv` for the client.

Webpack dotenv is like dotenv put has the extra advantange that any
variables that aren't used aren't exposed in the javascript.