Only access sequelize through the db
Use the sequelizer cli to `init` the model index.

The model index now loads the db client instead of the server.

All other sequalize models are loaded onto the `db` object.

Sequalize is now always accessed via the `db` object.

Switch to always using the `db` object everywhere.

Remove the models that were not auto-generated.

Change the name of the auto-generated models to be camelcase and remove
the timestamps from the definiation.

Setup associates via the associates method setup in the model index.