Add `yarn run dev-debug` to improve debugging
See the updated Local Development docs for an explaination of the added
power that this command gives you.

I added another runner because the new logging could add bunch of noise
other people don't like but really it could just be added to the
existing runner if people prefer.

I additionally looked into adding something like `express-winston` for
logging that could be used not just in for debugging but I just went
the route of just using the node debugger for now because for local dev
use it is faster to setup and provides about equavalint information.