Fix bug in database initialization script
When the query executed with the `psql` binary returns zero results, it
will print nothing to standard output, and the variable designed to
capture that output will contain an empty string.

If the `test` binary is invoked with a variable containing the empty
string, it will not recognize the value as a distinct argument and will
fail to apply comparison operators:

    $ bash -c 'x=""; [ $x == '1' ]'
    bash: line 0: [: ==: unary operator expected

Replace with the `[[` shell built-in to handle this condition
appropriately and to align with a similar pattern earlier in the same
script.

    $ bash -c 'x=""; [[ $x == '1' ]]'
    $