Report build failures during test import
Prior to this commit, the test-importing script would ignore errors in
the ARIA-AT project's build process. The script's subsequent behavior
was dependent on the particular failure, and if the problem was reported
at all, it would be in terms that were only indirectly related to the
root cause. For instance:

    Error found: Error: ENOENT: no such file or directory, scandir '/home/mike/aria-at-app/server/scripts/import-tests/tmp/build/tests'    [104/1957]
        at Object.readdirSync (fs.js:1021:3)
        at importTestPlanVersions (/home/mike/aria-at-app/server/scripts/import-tests/index.js:73:33)
    error Command failed with exit code 1.

Immediately exit the test-importing script if the ARIA-AT project's
build process fails for any reason.